+++ src/component/Interface.tsx
@@ -7,20 +7,18 @@ import Loading from './Loading';
 import { useAppState } from '../state';
 import animationState from '../state/animation';
 import loadingState from '../state/loading';
+import { useParams, ControlsOrientationValue } from '../state/params';
 import { FixTypeLater } from '../type/misc';
 import { fetchImagesFromManifest } from '../util/images';
 
 
-interface IInterfaceProps {
-  // TODO: Should this be nullable?
-  animationLocation: string | null;
-}
-
-const Interface: React.FC<IInterfaceProps> = (props) => {
+const Interface: React.FC = (props) => {
   const { state: appState, dispatch } = useAppState();
+  const { animationLocation, controlsOrientation, appId } = useParams();
 
   React.useEffect(
     () => {
+      // TODO: use animationLocation param instead
       if ( ! appState.animation.location ) {
         return;
       }
@@ -40,7 +38,7 @@ const Interface: React.FC<IInterfaceProps> = (props) => {
   );
   React.useEffect(
     () => {
-      if ( ! props.animationLocation ) {
+      if ( ! animationLocation ) {
         console.error(
           'No data found in "data-animation" attribute: ' +
           (appState.animation.location as string)
@@ -48,18 +46,22 @@ const Interface: React.FC<IInterfaceProps> = (props) => {
         dispatch(loadingState.actions.loaded());
         return;
       }
-      dispatch(animationState.actions.setLocation(props.animationLocation));
+      dispatch(animationState.actions.setLocation(animationLocation));
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [],
   );
 
-  return widgetContents(appState);
+  return widgetContents(appState, controlsOrientation, appId);
 }
 
 
 /* eslint-disable */
-const widgetContents = (state: FixTypeLater) => {
+const widgetContents = (
+  state: FixTypeLater,
+  controlsOrientation: ControlsOrientationValue,
+  appId: string,
+) => {
   if (state.loading) {
     return (
       <Loading />
@@ -67,7 +69,7 @@ const widgetContents = (state: FixTypeLater) => {
   } else if (state.animation.frames.length > 0) {
     return (
       <div className='animation-container'>
-        <ReactTooltip place={'bottom'} effect={'solid'} />
+        <ReactTooltip id={appId} place={controlsOrientation} effect={'solid'} />
         <Animation
           frames={state.animation.frames}
           playback={state.playback}